Izmjena na ljetnom zadatku
-popravak bug-a da netko ne moze brisati iz baze podataka preko GET metode (kopiranjem linka bez autorizacije)
-dodavanje 'Target="blank"' u linkove da otvara u novome prozoru
-dodavanje email-a i passworda ispod logina (da netko moze uci)
-popravljanje logiranja da petlja gleda samo email pa da svatko moze pristupiti samo s mailom
itd...

+++ brisanje.php
@@ -2,9 +2,15 @@
 
 if(!isset($_GET['sifra'])){
     header('location: index.php');
+    exit;
 }
 
 require_once 'konfiguracija.php';
+
+if(!isset($_SESSION['autoriziran'])){
+    header('location: index.php');
+    exit;
+    }
 
 $izraz=$veza->prepare('delete from riba where sifra=:sifra');
 $izraz->execute($_GET);
